% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{any_na_in_key}
\alias{any_na_in_key}
\title{Is there any NA in the key columns of the data.table?}
\usage{
any_na_in_key(x)
}
\arguments{
\item{x}{a \code{data.table} object}
}
\description{
Is there any NA in the key columns of the data.table?
}
\details{
It throws error when \code{x} doesn't have a key
}
