% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_no_na_in_key}
\alias{assert_no_na_in_key}
\title{Assert the key contains no NA}
\usage{
assert_no_na_in_key(x)
}
\arguments{
\item{x}{a \code{data.table} object}
}
\description{
Assert the key contains no NA
}
