% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_pk}
\alias{assert_pk}
\title{Assert the data.table has a primary key}
\usage{
assert_pk(x)
}
\arguments{
\item{x}{a \code{data.table} object}
}
\description{
The primary key here is similar to the concept in relational DB.
It means the key uniquely identifies each record in the table. It's an important
attribute when \code{x} is an information table, which is later quried / merged against
by other tables. A primary-keyed table can avoid duplicates in such cases.
}
\details{
Throws error unless \code{x} has a key, contains no \code{NA} values in key columns,
and is unique for key values.
}
