% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_unique}
\alias{assert_unique}
\title{Assert uniqueness of the data.table}
\usage{
assert_unique(x)
}
\arguments{
\item{x}{a \code{data.table} object}
}
\description{
Throws error when duplicates are found in the table
}
\details{
When \code{x} has a key, it only check the key columns.
}
