% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{get_duplicates}
\alias{get_duplicates}
\title{Find the duplicates}
\usage{
get_duplicates(x, only.key = FALSE, by = key(x))
}
\arguments{
\item{x}{a \code{data.table} object}

\item{only.key}{When \code{TRUE} and \code{x} has key, only the key columns are returned.}

\item{by}{The columns used for uniqueness checks. By default, it only relies on
the key columns of \code{x} to decide duplication. See \code{\link[data.table:duplicated]{data.table::duplicated()}}
for more info.}
}
\value{
The duplicated records of \code{x}.
}
\description{
Find the duplicates
}
\note{
When \code{x} has keys, the duplicateness is dependes on the key columns only.
}
