% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{is_pk}
\alias{is_pk}
\title{Does the data.table have a primary key?}
\usage{
is_pk(x)
}
\arguments{
\item{x}{a \code{data.table} object}
}
\value{
\code{TRUE} when the \code{data.table} object has a key,
has no \code{NA} in key columns and doesn't have duplicate key values.
Otherwise, it returns \code{FALSE}.
}
\description{
Does the data.table have a primary key?
}
