% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{is_unique}
\alias{is_unique}
\title{Is it a unique data.table?}
\usage{
is_unique(x)
}
\arguments{
\item{x}{a \code{data.table} object}
}
\value{
\code{TRUE} when there's no duplicate found in the table.
Otherwise, \code{FALSE} is returned.
}
\description{
Is it a unique data.table?
}
\details{
When \code{x} has a key, it only check the key columns.
}
