% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{NA_date_}
\alias{NA_date_}
\alias{NA_datetime_}
\title{The NA value with the Date / Date-Time Class}
\format{
An object of class \code{Date} of length 1.

An object of class \code{POSIXct} (inherits from \code{POSIXt}) of length 1.
}
\usage{
NA_date_

NA_datetime_
}
\description{
The NA value with the Date / Date-Time Class
}
\keyword{datasets}
