% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r_date}
\alias{r_date}
\title{Restore back to R Date}
\usage{
r_date(x)
}
\arguments{
\item{x}{the numeric values behind the Date value, which is the number of days
since 1970-01-01, with negative values for earlier dates.}
}
\description{
Calling R functions like \code{ifelse()} may accidentally convert a Date vector
to a numeric vector, due to losing the class attributes. This helper function
enables you to restore the numeric vector back to R Date.
}
\examples{
 r_date(18262)

}
