% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise_dt.R
\name{rowwise_dt}
\alias{rowwise_dt}
\title{Create a data.table rowwisely}
\usage{
rowwise_dt(...)
}
\arguments{
\item{...}{Arguments that defines the structure of a \code{data.table}.
The column names must be specified as \verb{col=} and may only appear before the data.
See the examples section for details.}
}
\value{
A \code{data.table} object. The default is for each column to return as a vector.
However, if any column has a length that is not one (e.g., \code{list(1, 2)}),
the whole column will be converted to a list column.
}
\description{
\code{rowwise_dt} allows the user to create a \code{data.table} object by
specifying a row-by-row layout. It's convenient and highly readable when
the table is small.
}
\note{
This function has been filed to data.table in
\href{https://github.com/Rdatatable/data.table/pull/4291}{PR#4291}, where the function
name is \code{rowwiseDT}.
}
\examples{
 rowwise_dt(
   A=, B=, C=,
   1, "a", 2:3,
   2, "b", list(5)
 )
}
