% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_taken}
\alias{time_taken}
\title{Convinient Time Taken}
\usage{
time_taken(reset = FALSE)
}
\arguments{
\item{reset}{When \code{TRUE}, it resets the clock.}
}
\description{
Compared to \code{\link[data.table:timetaken]{data.table::timetaken()}}, this version doesn't require the user
to store the value of \code{\link[base:proc.time]{base::proc.time()}} explicitly. Instead, the first call in a
new session will store the value of proc.time automatically. Or you can call the
function with param \code{TRUE} to reset the stored proc.time value, afterwards.
}
\examples{
 time_taken()
 Sys.sleep(1)
 cat("Finished in",time_taken(),"\n")
}
